requirement.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
bidict==0.23.1
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
braceexpand==0.1.7
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
contexttimer==0.3.3
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
decorator==5.1.1
decord==0.6.0
distlib==0.3.8
dlib==19.24.2
docutils==0.20.1
easyocr==1.7.1
einops==0.7.0
executing==2.0.1
face-recognition==1.3.0
face-recognition-models==0.3.0
fairscale==0.4.4
fastapi==0.109.2
ffmpy==0.3.2
filelock==3.13.1
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.1.3
gitdb==4.0.11
GitPython==3.1.42
gradio==4.19.2
gradio_client==0.10.1
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.20.3
identify==2.5.35
idna==3.6
ifaddr==0.2.0
ImageHash==4.3.1
imageio==2.34.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
iopath==0.1.10
ipython==8.22.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kaggle==1.6.6
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
networkx==3.2.1
nicegui==1.4.16
ninja==1.11.1.1
nodeenv==1.8.0
numpy==1.26.4
omegaconf==2.3.0
opencv-python==4.9.0.80
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
orjson==3.9.15
outcome==1.3.0.post0
packaging==23.2
pandas==2.2.1
parso==0.8.3
pillow==10.2.0
platformdirs==4.2.0
plotly==5.19.0
portalocker==2.8.2
pre-commit==3.6.2
preshed==3.0.9
prompt-toolkit==3.0.43
protobuf==4.25.3
pscript==0.7.7
pure-eval==0.2.2
pyarrow==15.0.0
pyclipper==1.3.0.post5
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser==2.21
pydantic==2.6.2
pydantic_core==2.16.3
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.17.2
pyparsing==3.1.1
PySocks==1.7.1
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-magic==0.4.27
python-multipart==0.0.9
python-slugify==8.0.4
python-socketio==5.11.1
pytz==2024.1
PyWavelets==1.5.0
pywin32==306
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.18.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
ruff==0.2.2
salesforce-lavis==1.0.2
scikit-image==0.22.0
scipy==1.12.0
selenium==4.9.0
semantic-version==2.10.0
sentencepiece==0.2.0
shapely==2.0.3
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.36.3
streamlit==1.31.1
sympy==1.12
tenacity==8.2.3
text-unidecode==1.3
thinc==8.2.3
tifffile==2024.2.12
timm==0.4.12
tokenizers==0.13.3
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.1+cu121
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.26.1
trio==0.24.0
trio-websocket==0.11.1
typer==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
urllib3==1.26.18
uvicorn==0.27.1
validators==0.22.0
vbuild==0.8.2
virtualenv==20.25.1
wasabi==1.1.2
watchdog==4.0.0
watchfiles==0.21.0
wcwidth==0.2.13
weasel==0.3.4
webdataset==0.2.86
webdriver-manager==4.0.1
webencodings==0.5.1
websockets==11.0.3
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
